Project 4 - Kmeans and EM Algorithm
Files for the first two points of the fourth project. For the EM algorithm, both the E-Step and the M-step are wrong. The E-step is wrong because I was using just a regular gaussian equation to get the likelihood, the M-step on the other hand is almost ok, just the variances are a little off. I need to check both of those issues.